Add fallback to procfs after try netlink
